@@ -3,13 +3,10 @@
 #include <iostream>
 
 #include "boost/algorithm/string.hpp"
-#include "boost/filesystem/operations.hpp"
 
 #include "json/json.h"
 
 #include "book_json.h"
-
-namespace bfs = boost::filesystem;
 
 BookClient::BookClient(const std::string& url, int timeout)
     : url_(url), session_(timeout) {
@@ -127,7 +124,7 @@ bool BookClient::Delete(const std::string& id) {
   }
 }
 
-bool BookClient::GetPhoto(const std::string& id, const bfs::path& path) {
+bool BookClient::GetPhoto(const std::string& id, const webcc::fs::path& path) {
   try {
     auto r = session_.Send(WEBCC_GET(url_).
                            Path("books").Path(id).Path("photo")(),
@@ -147,7 +144,7 @@ bool BookClient::GetPhoto(const std::string& id, const bfs::path& path) {
   }
 }
 
-bool BookClient::SetPhoto(const std::string& id, const bfs::path& path) {
+bool BookClient::SetPhoto(const std::string& id, const webcc::fs::path& path) {
   try {
     if (!CheckPhoto(path)) {
       return false;
@@ -169,12 +166,12 @@ bool BookClient::SetPhoto(const std::string& id, const bfs::path& path) {
   }
 }
 
-bool BookClient::CheckPhoto(const bfs::path& photo) {
+bool BookClient::CheckPhoto(const webcc::fs::path& photo) {
   if (photo.empty()) {
     return false;
   }
 
-  if (!bfs::is_regular_file(photo) || !bfs::exists(photo)) {
+  if (!webcc::fs::is_regular_file(photo) || !webcc::fs::exists(photo)) {
     return false;
   }
 
