@@ -4,8 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "boost/filesystem/path.hpp"
-
+#include "webcc/fs.h"
 #include "webcc/request.h"
 #include "webcc/response.h"
 
@@ -96,8 +95,7 @@ public:
 
   // Use the file content as body.
   // NOTE: Error::kFileError might be thrown.
-  ResponseBuilder& File(const boost::filesystem::path& path,
-                        bool infer_media_type = true,
+  ResponseBuilder& File(const fs::path& path, bool infer_media_type = true,
                         std::size_t chunk_size = 1024);
 
   ResponseBuilder& Header(const std::string& key, const std::string& value) {
